Backend/servers/03 catalogues/pyproject.toml
[tool.poetry]
name = "catalogues"
version = "0.1.0"
description = ""
authors = ["1twalters01 <1twalters01@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
run = "src.main:application"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.1"
uvicorn = "^0.30.6"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.32"
psycopg2-binary = "^2.9.9"
python-dateutil = "^2.9.0.post0"
pyjwt = "^2.9.0"
protobuf = "^5.28.2"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
pytest = "^8.3.2"

[[tool.poetry.packages]]
include = "src"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
